+++ src/api/subscriptions/infrastructure/persistence/entities/plan.entity.spec.ts
